skeleton of split design
