+++ lib/hardware_services/inventory_drivers/QuadsNativeInventoryDriver.py
@@ -12,9 +12,9 @@ import logging
 from subprocess import call
 from subprocess import check_call
 
-from hardware_services.hardware_service import HardwareService
+from hardware_services.inventory_service import InventoryService
 
-class JuniperDriver(HardwareService):
+class QuadsNativeInventoryDriver(InventoryService):
 
     def update_cloud(self, quadsinstance, **kwargs):
         if kwargs['description'] is None:
@@ -99,33 +99,6 @@ class JuniperDriver(HardwareService):
 
         return
 
-    def move_hosts(self, quadsinstance, **kwargs):
-        # move a host
-        for h in sorted(quadsinstance.quads.hosts.data.iterkeys()):
-            default_cloud, current_cloud, current_override = quadsinstance._quads_find_current(h, kwargs['datearg'])
-            if not os.path.isfile(kwargs['statedir'] + "/" + h):
-                try:
-                    stream = open(kwargs['statedir'] + "/" + h, 'w')
-                    stream.write(current_cloud + '\n')
-                    stream.close()
-                except Exception, ex:
-                    quadsinstance.logger.error("There was a problem with your file %s" % ex)
-            else:
-                stream = open(kwargs['statedir'] + "/" + h, 'r')
-                current_state = stream.readline().rstrip()
-                stream.close()
-                if current_state != current_cloud:
-                    quadsinstance.logger.info("Moving " + h + " from " + current_state + " to " + current_cloud)
-                    if not kwargs['dryrun']:
-                        try:
-                            check_call([kwargs['movecommand'], h, current_state, current_cloud])
-                        except Exception, ex:
-                            quadsinstance.logger.error("Move command failed: %s" % ex)
-                            exit(1)
-                        stream = open(kwargs['statedir'] + "/" + h, 'w')
-                        stream.write(current_cloud + '\n')
-                        stream.close()
-        return
 
     def list_clouds(self,quadsinstance):
         quadsinstance.quads.clouds.cloud_list()
